make halfWindowSize an Int

+++ src/main/scala/org/hammerlab/guacamole/distributed/WindowFlatMapUtils.scala
@@ -42,7 +42,7 @@ object WindowFlatMapUtils {
   def windowFlatMapWithState[R <: ReferenceRegion: ClassTag, T: ClassTag, S](regionRDDs: PerSample[RDD[R]],
                                                                              lociPartitions: LociPartitioning,
                                                                              skipEmpty: Boolean,
-                                                                             halfWindowSize: Long,
+                                                                             halfWindowSize: Int,
                                                                              initialState: S,
                                                                              function: (S, PerSample[SlidingWindow[R]]) => (S, Iterator[T])): RDD[T] = {
     windowTaskFlatMapMultipleRDDs(
@@ -87,7 +87,7 @@ object WindowFlatMapUtils {
   def windowFoldLoci[R <: ReferenceRegion: ClassTag, T: ClassTag](regionRDDs: PerSample[RDD[R]],
                                                                   lociPartitions: LociPartitioning,
                                                                   skipEmpty: Boolean,
-                                                                  halfWindowSize: Long,
+                                                                  halfWindowSize: Int,
                                                                   initialValue: T,
                                                                   aggFunction: (T, PerSample[SlidingWindow[R]]) => T): RDD[T] = {
     windowTaskFlatMapMultipleRDDs(
@@ -140,7 +140,7 @@ object WindowFlatMapUtils {
   private[distributed] def windowTaskFlatMapMultipleRDDs[R <: ReferenceRegion: ClassTag, T: ClassTag](
     regionRDDs: PerSample[RDD[R]],
     lociPartitions: LociPartitioning,
-    halfWindowSize: Long,
+    halfWindowSize: Int,
     function: (Long, LociSet, PerSample[Iterator[R]]) => Iterator[T]): RDD[T] = {
 
     val numRDDs = regionRDDs.length
@@ -229,7 +229,7 @@ object WindowFlatMapUtils {
   def collectByContig[R <: ReferenceRegion: ClassTag, T: ClassTag](
     taskRegionsPerSample: PerSample[Iterator[R]],
     taskLoci: LociSet,
-    halfWindowSize: Long,
+    halfWindowSize: Int,
     generateFromWindows: (Contig, PerSample[SlidingWindow[R]]) => Iterator[T]): Iterator[T] = {
 
     val regionSplitByContigPerSample: PerSample[RegionsByContig[R]] = taskRegionsPerSample.map(new RegionsByContig(_))
